backend/pyproject.toml
[tool.poetry]
name = "kisaweb"
version = "0.1.0"
description = ""
authors = ["Jonathan Willianto <jowillianto@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
alabaster = "^0.7.13"
appnope = "^0.1.3"
asgiref = "^3.7.2"
astroid = "^3.0.1"
asttokens = "^2.4.1"
babel = "^2.13.1"
backcall = "^0.2.0"
black = "^23.11.0"
certifi = "^2023.11.17"
cfgv = "^3.4.0"
charset-normalizer = "^3.3.2"
click = "^8.1.7"
coverage = "^7.3.2"
decorator = "^5.1.1"
dill = "^0.3.7"
distlib = "^0.3.7"
django = "^4.2.7"
django-admin-sortable = "^2.3"
django-appconf = "^1.0.5"
django-cors-headers = "^4.3.1"
django-compressor = "^4.4"
django-crispy-forms = "^2.1"
django-docs = "^0.3.3"
django-filter = "^23.3"
django-libsass = "^0.9"
django-maintenance-mode = "^0.19.0"
django-pagination-bootstrap = "^2.4.6"
django-phone-field = "^1.8.1"
django-stubs = "^4.2.6"
django-tinymce = "^3.6.1"
djangorestframework = "^3.14.0"
exceptiongroup = "^1.1.3"
executing = "^2.0.1"
filelock = "^3.13.1"
gunicorn = "^21.2.0"
identify = "^2.5.31"
idna = "^3.4"
imagesize = "^1.4.1"
importlib-metadata = "^6.8.0"
iniconfig = "^2.0.0"
ipython = "^8.17.2"
isort = "^5.12.0"
jedi = "^0.19.1"
jinja2 = "^3.1.2"
jsonschema = "^4.20.0"
lazy-object-proxy = "^1.9.0"
libsass = "^0.22.0"
markdown = "^3.4.4"
markupsafe = "^2.1.3"
matplotlib-inline = "^0.1.6"
mccabe = "^0.7.0"
mypy = "^1.7.0"
mypy-extensions = "^1.0.0"
nodeenv = "^1.8.0"
packaging = "^23.2"
parso = "^0.8.3"
pathspec = "^0.11.2"
pexpect = "^4.8.0"
pickleshare = "^0.7.5"
pillow = "^10.1.0"
pluggy = "^1.3.0"
pre-commit = "^3.5.0"
prompt-toolkit = "^3.0.41"
ptyprocess = "^0.7.0"
pure-eval = "^0.2.2"
pycryptodome = "^3.19.0"
pygments = "^2.16.1"
pylint = "^3.0.2"
pylint-django = "^2.5.5"
pylint-plugin-utils = "^0.8.2"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-django = "^4.7.0"
python-dotenv = "^1.0.0"
pytz = "^2023.3.post1"
pyyaml = "^6.0.1"
requests = "^2.31.0"
setproctitle = "^1.3.3"
six = "^1.16.0"
snowballstemmer = "^2.2.0"
sphinx = "^7.2.6"
sphinx-rtd-theme = "^1.3.0"
sphinxcontrib-applehelp = "^1.0.7"
sphinxcontrib-devhelp = "^1.0.5"
sphinxcontrib-htmlhelp = "^2.0.4"
sphinxcontrib-jquery = "^4.1"
sphinxcontrib-jsmath = "^1.0.1"
sphinxcontrib-qthelp = "^1.0.6"
sphinxcontrib-serializinghtml = "^1.1.9"
sqlparse = "^0.4.4"
stack-data = "^0.6.3"
tomli = "^2.0.1"
tomlkit = "^0.12.3"
traitlets = "^5.13.0"
typed-ast = "^1.5.5"
urllib3 = "^2.1.0"
virtualenv = "^20.24.6"
wcwidth = "^0.2.10"
wrapt = "^1.16.0"
zipp = "^3.17.0"
psycopg2-binary = "^2.9.9"
pypharmaco = {url = "https://github.com/Calici/pypharmaco/releases/download/v0.0.1/pypharmaco-0.0.1-py3-none-any.whl"}
google-api-python-client = "^2.113.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
